feat : added complete implementation of proving job
